+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added full search history support on the "Search Dialog".
 - Added new search setting "history", which is used to state the existing search history (is updated when new entries are added).
 - Added single day (in month) view toggle support (simply double click on the day name header you want to view).
+- Added "Cut" support across events (like copy, but instead moves the event to the desired date).
 - 
 
 ### Version 1.7.5: